Djperrefort/scratch (#27)
This PR rebuilds our analysis code from scratch. Given the number of problems we were having with the original code, it seemed easiest to simply start over. Benefits and changes of this PR include:

1. Preliminary results from the code in this branch look much more promising
2. Significantly higher test coverage
3. A simpler, smaller codebase
4. Full HTML documentation and notes
5. Updated and simplified notebooks (that actually run!)

There is some missing functionality that is removed by this PR. Most notably the ability to manually specify priors (which will be added in later) and a few notebooks that haven't been ported yet (see the *notebooks/old_notebooks* directory).