Switch to hsiao (#33)
This pull request switches from using the Salt2 model to the Hsiao model. It also introduces the use of band-by-band fitting to more closely resemble the approach of Sifto. The Hsiao model is a customized version of the Hsiao 3.0 model built into SNCosmo. It differs only in the addition of a stretch x1 parameter. The new model is registered with sncosmo under the name hsiao_x1.